Alteração Braba

+++ src/main/java/br/edu/ufersa/pizzaria/Michelangelo/domain/entity/Additional.java
@@ -1,5 +1,6 @@
 package br.edu.ufersa.pizzaria.Michelangelo.domain.entity;
 
+import java.math.BigDecimal;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -25,11 +26,11 @@ public class Additional {
 
   @PositiveOrZero
   @Column(precision = 5, scale = 2)
-  private float price;
+  private BigDecimal price;
 
   private String image;
 
-  public Additional(String name, String description, float price, String image) {
+  public Additional(String name, String description, BigDecimal price, String image) {
     this.name = name;
     this.description = description;
     this.price = price;
@@ -67,14 +68,14 @@ public class Additional {
   /**
    * @return float return the price
    */
-  public float getPrice() {
+  public BigDecimal getPrice() {
     return price;
   }
 
   /**
    * @param price the price to set
    */
-  public void setPrice(float price) {
+  public void setPrice(BigDecimal price) {
     this.price = price;
   }
